# apps/api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/app": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "appTag"
        ]
      }
    },
    "/auth/verifyCode": {
      "get": {
        "operationId": "AuthController_verifyCode",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取验证码",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomResponseDto_VerifyCodeDto"
                }
              }
            }
          }
        },
        "summary": "获取验证码",
        "tags": [
          "认证模块"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "账号密码登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomResponseDto_UserInfoDto"
                }
              }
            }
          }
        },
        "summary": "账号密码登录",
        "tags": [
          "认证模块"
        ]
      }
    },
    "/auth/current": {
      "get": {
        "operationId": "AuthController_current",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取当前登录用户",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomResponseDto_SysUserModel"
                }
              }
            }
          }
        },
        "summary": "获取当前登录用户",
        "tags": [
          "认证模块"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "注册用户",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomResponseDto_UserInfoDto"
                }
              }
            }
          }
        },
        "summary": "注册用户",
        "tags": [
          "认证模块"
        ]
      }
    },
    "/api/sys/user": {
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建用户",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomResponseDto_SysUserModel"
                }
              }
            }
          }
        },
        "summary": "创建用户",
        "tags": [
          "用户管理"
        ]
      },
      "get": {
        "operationId": "UserController_findAll",
        "parameters": [
          {
            "name": "pageNo",
            "required": false,
            "in": "query",
            "description": "当前页码",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页数据条数",
            "schema": {
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询所有用户",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomResponseDto_PaginationResultDto"
                }
              }
            }
          }
        },
        "summary": "查询所有用户",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/sys/user/{id}": {
      "get": {
        "operationId": "UserController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询用户详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomResponseDto_SysUserModel"
                }
              }
            }
          }
        },
        "summary": "查询用户详情",
        "tags": [
          "用户管理"
        ]
      },
      "patch": {
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新用户",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomResponseDto_SysUserModel"
                }
              }
            }
          }
        },
        "summary": "更新用户",
        "tags": [
          "用户管理"
        ]
      },
      "delete": {
        "operationId": "UserController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除用户",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomResponseDto_SysUserModel"
                }
              }
            }
          }
        },
        "summary": "删除用户",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/sys/role": {
      "post": {
        "operationId": "RoleController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "角色管理"
        ]
      },
      "get": {
        "operationId": "RoleController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "角色管理"
        ]
      }
    },
    "/api/sys/role/{id}": {
      "get": {
        "operationId": "RoleController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "角色管理"
        ]
      },
      "patch": {
        "operationId": "RoleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "角色管理"
        ]
      },
      "delete": {
        "operationId": "RoleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "角色管理"
        ]
      }
    },
    "/sys/menu/{id}/detail": {
      "get": {
        "description": "获取菜单详情",
        "operationId": "MenuController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomResponseDto_SysMenuModel"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "系统菜单"
        ]
      }
    },
    "/sys/menu/list": {
      "get": {
        "description": "获取菜单列表",
        "operationId": "MenuController_findList",
        "parameters": [
          {
            "name": "pageNo",
            "required": false,
            "in": "query",
            "description": "当前页码",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页数据条数",
            "schema": {
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomResponseDto_PaginationResultDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "系统菜单"
        ]
      }
    },
    "/sys/menu": {
      "post": {
        "description": "创建菜单",
        "operationId": "MenuController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMenuDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomResponseDto_SysMenuModel"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "系统菜单"
        ]
      }
    },
    "/sys/menu/{id}": {
      "put": {
        "description": "更新菜单",
        "operationId": "MenuController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMenuDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomResponseDto_SysMenuModel"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "系统菜单"
        ]
      },
      "delete": {
        "description": "删除菜单",
        "operationId": "MenuController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomResponseDto_SysMenuModel"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "系统菜单"
        ]
      }
    },
    "/test/getToken": {
      "get": {
        "operationId": "TestController_getToken",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "获取测试token",
        "tags": [
          "test"
        ]
      }
    },
    "/test/checkPermission": {
      "get": {
        "operationId": "TestController_checkPermission",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "检查权限",
        "tags": [
          "test"
        ]
      }
    },
    "/test/test": {
      "post": {
        "operationId": "TestController_test",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "测试接口",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomResponseDto_TestDto"
                }
              }
            }
          }
        },
        "summary": "测试接口",
        "tags": [
          "test"
        ]
      }
    }
  },
  "info": {
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "in": "header",
        "description": "请输入 JWT 令牌"
      }
    },
    "schemas": {
      "VerifyCodeDto": {
        "type": "object",
        "properties": {
          "verifyId": {
            "type": "string",
            "description": "验证id"
          }
        }
      },
      "CustomResponseDto_VerifyCodeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "状态码"
          },
          "msg": {
            "type": "string",
            "description": "消息"
          },
          "now": {
            "type": "number",
            "description": "请求时间"
          },
          "data": {
            "$ref": "#/components/schemas/VerifyCodeDto"
          }
        },
        "required": [
          "data"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "description": "密码"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "UserInfoDto": {
        "type": "object",
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer",
            "description": "主键ID"
          },
          "createTime": {
            "format": "date-time",
            "type": "string"
          },
          "updateUser": {
            "type": "object",
            "description": "修改人",
            "nullable": true
          },
          "updateTime": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "gender": {
            "type": "number"
          },
          "avatar": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "number"
          },
          "isSystem": {
            "type": "boolean"
          },
          "pwdResetTime": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "deptId": {
            "format": "int64",
            "type": "integer",
            "description": "部门ID"
          },
          "token": {
            "type": "string",
            "description": "密钥"
          },
          "phone": {
            "type": "string",
            "description": "手机号"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          }
        },
        "required": [
          "id",
          "createTime",
          "username",
          "nickname",
          "gender",
          "status",
          "isSystem",
          "deptId"
        ]
      },
      "CustomResponseDto_UserInfoDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "状态码"
          },
          "msg": {
            "type": "string",
            "description": "消息"
          },
          "now": {
            "type": "number",
            "description": "请求时间"
          },
          "data": {
            "$ref": "#/components/schemas/UserInfoDto"
          }
        },
        "required": [
          "data"
        ]
      },
      "SysUserModel": {
        "type": "object",
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer",
            "description": "主键ID"
          },
          "createUser": {
            "type": "object",
            "description": "创建人",
            "nullable": true
          },
          "createTime": {
            "format": "date-time",
            "type": "string"
          },
          "updateUser": {
            "type": "object",
            "description": "修改人",
            "nullable": true
          },
          "updateTime": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "deleted": {
            "format": "int64",
            "type": "integer",
            "description": "是否已删除"
          },
          "username": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "number"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "avatar": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "number"
          },
          "isSystem": {
            "type": "boolean"
          },
          "pwdResetTime": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "deptId": {
            "format": "int64",
            "type": "integer",
            "description": "部门ID"
          }
        },
        "required": [
          "id",
          "createTime",
          "deleted",
          "username",
          "nickname",
          "gender",
          "status",
          "isSystem",
          "deptId"
        ]
      },
      "CustomResponseDto_SysUserModel": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "状态码"
          },
          "msg": {
            "type": "string",
            "description": "消息"
          },
          "now": {
            "type": "number",
            "description": "请求时间"
          },
          "data": {
            "$ref": "#/components/schemas/SysUserModel"
          }
        },
        "required": [
          "data"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "description": "密码"
          }
        }
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "phone": {
            "type": "string",
            "description": "手机号"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          }
        }
      },
      "PaginationResultDto": {
        "type": "object",
        "properties": {
          "pageNo": {
            "type": "number",
            "description": "当前页码",
            "default": 1
          },
          "pageSize": {
            "type": "number",
            "description": "每页数据条数",
            "default": 10
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SysUserModel"
            }
          },
          "total": {
            "type": "number",
            "description": "总条数"
          }
        }
      },
      "CustomResponseDto_PaginationResultDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "状态码"
          },
          "msg": {
            "type": "string",
            "description": "消息"
          },
          "now": {
            "type": "number",
            "description": "请求时间"
          },
          "data": {
            "$ref": "#/components/schemas/PaginationResultDto"
          }
        },
        "required": [
          "data"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "phone": {
            "type": "string",
            "description": "手机号"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          }
        }
      },
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "角色名称",
            "maxLength": 255,
            "example": "管理员"
          },
          "description": {
            "type": "string",
            "description": "角色描述",
            "maxLength": 500,
            "example": "系统管理员角色"
          }
        }
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "角色名称",
            "maxLength": 255,
            "example": "管理员"
          },
          "description": {
            "type": "string",
            "description": "角色描述",
            "maxLength": 500,
            "example": "系统管理员角色"
          }
        }
      },
      "SysMenuModel": {
        "type": "object",
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer",
            "description": "主键ID"
          },
          "createUser": {
            "format": "int64",
            "type": "integer",
            "description": "创建人"
          },
          "createTime": {
            "format": "date-time",
            "type": "string"
          },
          "updateUser": {
            "type": "object",
            "description": "修改人",
            "nullable": true
          },
          "updateTime": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "deleted": {
            "format": "int64",
            "type": "integer",
            "description": "是否已删除"
          },
          "status": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "parentId": {
            "format": "int64",
            "type": "integer",
            "description": "上级菜单ID"
          },
          "type": {
            "type": "number"
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "component": {
            "type": "string",
            "nullable": true
          },
          "redirect": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "isExternal": {
            "type": "boolean",
            "nullable": true
          },
          "isCache": {
            "type": "boolean",
            "nullable": true
          },
          "isHidden": {
            "type": "boolean",
            "nullable": true
          },
          "permission": {
            "type": "string",
            "nullable": true
          },
          "sort": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "createUser",
          "createTime",
          "deleted",
          "status",
          "title",
          "parentId",
          "type",
          "sort"
        ]
      },
      "CustomResponseDto_SysMenuModel": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "状态码"
          },
          "msg": {
            "type": "string",
            "description": "消息"
          },
          "now": {
            "type": "number",
            "description": "请求时间"
          },
          "data": {
            "$ref": "#/components/schemas/SysMenuModel"
          }
        },
        "required": [
          "data"
        ]
      },
      "CreateMenuDto": {
        "type": "object",
        "properties": {}
      },
      "TestDto": {
        "type": "object",
        "properties": {
          "TestField": {
            "type": "string",
            "description": "测试字段"
          },
          "test_field2": {
            "type": "string",
            "description": "hello"
          }
        }
      },
      "CustomResponseDto_TestDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "状态码"
          },
          "msg": {
            "type": "string",
            "description": "消息"
          },
          "now": {
            "type": "number",
            "description": "请求时间"
          },
          "data": {
            "$ref": "#/components/schemas/TestDto"
          }
        },
        "required": [
          "data"
        ]
      }
    }
  },
  "security": [
    {
      "Authorization": []
    }
  ]
}